(I)DAL changes CSVWriter/CSVReader changes ZeitErfassungVier changes
git-svn-id: file:///home/node/svn/project_se/goran/EPU_Office@146 bde5ac68-c2e4-480a-bd7c-0859b87ef666

+++ src/utils/csv/CSVReader.java
@@ -99,7 +99,7 @@ public class CSVReader {
                         counter++;
                     }
                     if (counter == 3) {
-                        logEintrag.setStunden(Integer.parseInt(st.nextToken()));
+                        logEintrag.setStunden((Integer)Integer.parseInt(st.nextToken()));
                         System.out.println(counter + " Stunden  " + logEintrag.getStunden());
                         counter++;
                     }
